Moved datadownloader test to long running folder
