+++ js/src/dataDownloader/downloadData.spec.ts
@@ -20,30 +20,4 @@ describe('downloadData', () => {
                 expect(result.endsWith('.config/openenf')).toBe(true);
         }
     })
-    it('can download data file', (done) => {
-        const filePath = path.join(getENFDataDirectory(),"xx.freqdb");
-        if (fs.existsSync(filePath)) {
-            fs.rmSync(filePath);
-        }
-        downloadData("XX").then(() => {
-            expect(fs.existsSync(filePath)).toBe(true);
-            const fileHash = crypto.createHash('md5');
-            const stream = fs.createReadStream(filePath);
-
-            stream.on('data', function(data) {
-                fileHash.update(data);
-            });
-
-            stream.on('end', function() {
-                const md5sum = fileHash.digest('hex');
-                expect(md5sum).toBe('ddc138b743874940c9577e03df5b21db')
-                fs.rmSync(filePath);
-                done();
-            });
-
-            stream.on('error', function(err) {
-                console.error(`Error reading file: ${err}`);
-            });
-        })
-    })
 })
